tele-vue-back/requirements.txt
fastapi[all]
passlib[bcrypt]
psycopg2
pyjwt
sqlalchemy
mimesis
